+ Lambdas implemented
However, type population of lambdas is not DEEP. This lambda `i -> i * 2` will not have its type signature concretized at compile time:

`(int -> int)[] fs = [ i -> i * 2 ];`

This is because the interpreter does not search into collection data structures to look for function types to use as signature suppliers for RHS lambdas.

Happy to accept this for now, as correct behaviour should not be compromised.